move server_api to api
