+++ client/lib/src/lib.rs
@@ -6,7 +6,6 @@ pub mod config;
 pub mod sc_manager;
 pub mod wr_manager;
 pub mod device;
-pub mod server_api;
 pub mod api;
 pub mod flutter_api;
 pub mod client_manager;
